==> notebooks/03_feature_engineering_old.ipynb <==
# Swiggy Restaurant Recommendation System  
## Feature Engineering and Encoding

### Objective
The objective of this notebook is to prepare the cleaned restaurant dataset
for recommendation modeling by encoding categorical features and ensuring
all values are numerical.


import pandas as pd
import numpy as np

from sklearn.preprocessing import OneHotEncoder
import pickle


# Load cleaned dataset
df = pd.read_csv("../data/processed/cleaned_data.csv")

# Preview dataset
df.head()


df.shape

### Feature Selection

The following features are selected for recommendation modeling:

- Categorical Features: city, cuisine
- Numerical Features: rating, rating_count, cost


# Selecting relevant features
categorical_features = ['city', 'cuisine']
numerical_features = ['rating', 'rating_count', 'cost']

df_selected = df[categorical_features + numerical_features]
df_selected.head()


# Initialize One-Hot Encoder
encoder = OneHotEncoder(sparse_output=False, handle_unknown='ignore')



# Fit and transform categorical features
encoded_cat = encoder.fit_transform(df_selected[categorical_features])

# Get encoded feature names
encoded_cat_columns = encoder.get_feature_names_out(categorical_features)

# Convert encoded data to DataFrame
encoded_cat_df = pd.DataFrame(encoded_cat, columns=encoded_cat_columns)

encoded_cat_df.head()


# Reset index to maintain alignment
encoded_cat_df.reset_index(drop=True, inplace=True)
df_selected.reset_index(drop=True, inplace=True)


# Combine numerical and encoded categorical features
final_df = pd.concat(
    [df_selected[numerical_features], encoded_cat_df],
    axis=1
)

final_df.head()


final_df.shape


# Save encoded dataset
final_df.to_csv("../data/processed/encoded_data.csv", index=False)


# Save encoder for later use in Streamlit
with open("../models/encoder.pkl", "wb") as file:
    pickle.dump(encoder, file)
